+++ buildVars.py
@@ -14,90 +14,73 @@ from site_scons.site_tools.NVDATool.utils import _
 # Add-on information variables
 addon_info = AddonInfo(
 	# add-on Name/identifier, internal for NVDA
-	addon_name="addonTemplate",
+	addon_name="Open_Bible",
 	# Add-on summary/title, usually the user visible name of the add-on
 	# Translators: Summary/title for this add-on
 	# to be shown on installation and add-on information found in add-on store
-	addon_summary=_("Add-on user visible name"),
+	addon_summary=_("Open Bible"),
 	# Add-on description
 	# Translators: Long description to be shown for this add-on on add-on information from add-on store
-	addon_description=_("""Description for the add-on.
-It can span multiple lines."""),
+	addon_description=_(
+		"""Este add-on permite a leitura da Bíblia Sagrada no NVDA em várias versões (arquivos JSON).
+		Você pode navegar por livro, capítulo e versículo, pesquisar, favoritar, anotar e copiar/compartilhar versículos."""
+	),
 	# version
-	addon_version="x.y",
+	addon_version="2026.01.31",
 	# Brief changelog for this version
 	# Translators: what's new content for the add-on version to be shown in the add-on store
-	addon_changelog=_("""Changelog for the add-on version.
-It can span multiple lines."""),
+	addon_changelog=_(
+		"""Primeira versão publicada no GitHub com estrutura padronizada do template oficial.
+		Inclui o atalho Ctrl+Alt+B para abrir a janela principal."""
+	),
 	# Author(s)
-	addon_author="name <name@domain.com>",
+	addon_author="Leandro Souza",
 	# URL for the add-on documentation support
-	addon_url=None,
+	addon_url="https://github.com/leandro-sds/Open-Bible",
 	# URL for the add-on repository where the source code can be found
-	addon_sourceURL=None,
+	addon_sourceURL="https://github.com/leandro-sds/Open-Bible",
 	# Documentation file name
 	addon_docFileName="readme.html",
 	# Minimum NVDA version supported (e.g. "2019.3.0", minor version is optional)
-	addon_minimumNVDAVersion=None,
+	addon_minimumNVDAVersion="2024.1",
 	# Last NVDA version supported/tested (e.g. "2024.4.0", ideally more recent than minimum version)
-	addon_lastTestedNVDAVersion=None,
-	# Add-on update channel (default is None, denoting stable releases,
-	# and for development releases, use "dev".)
+	addon_lastTestedNVDAVersion="2025.2.3",
+	# Add-on update channel (default is None, denoting stable releases, and for development releases, use "dev".)
 	# Do not change unless you know what you are doing!
 	addon_updateChannel=None,
 	# Add-on license such as GPL 2
-	addon_license=None,
+	addon_license="GPL-2.0-only",
 	# URL for the license document the ad-on is licensed under
 	addon_licenseURL=None,
 )
 
 # Define the python files that are the sources of your add-on.
-# You can either list every file (using ""/") as a path separator,
-# or use glob expressions.
-# For example to include all files with a ".py" extension from the "globalPlugins" dir of your add-on
-# the list can be written as follows:
-# pythonSources = ["addon/globalPlugins/*.py"]
-# For more information on SCons Glob expressions please take a look at:
-# https://scons.org/doc/production/HTML/scons-user/apd.html
-pythonSources: list[str] = []
+# Paths are relative to the repository root. Use "/" as separator.
+pythonSources: list[str] = [
+	"addon/globalPlugins/open_bible.py",
+]
 
 # Files that contain strings for translation. Usually your python sources
 i18nSources: list[str] = pythonSources + ["buildVars.py"]
 
 # Files that will be ignored when building the nvda-addon file
 # Paths are relative to the addon directory, not to the root directory of your addon sources.
-# You can either list every file (using ""/") as a path separator,
-# or use glob expressions.
 excludedFiles: list[str] = []
 
 # Base language for the NVDA add-on
-# If your add-on is written in a language other than english, modify this variable.
-# For example, set baseLanguage to "es" if your add-on is primarily written in spanish.
-# You must also edit .gitignore file to specify base language files to be ignored.
-baseLanguage: str = "en"
+baseLanguage: str = "pt"
 
 # Markdown extensions for add-on documentation
-# Most add-ons do not require additional Markdown extensions.
-# If you need to add support for markup such as tables, fill out the below list.
-# Extensions string must be of the form "markdown.extensions.extensionName"
-# e.g. "markdown.extensions.tables" to add tables.
 markdownExtensions: list[str] = []
 
 # Custom braille translation tables
-# If your add-on includes custom braille tables (most will not), fill out this dictionary.
-# Each key is a dictionary named according to braille table file name,
-# with keys inside recording the following attributes:
-# displayName (name of the table shown to users and translatable),
-# contracted (contracted (True) or uncontracted (False) braille code),
-# output (shown in output table list),
-# input (shown in input table list).
 brailleTables: BrailleTables = {}
 
-# Custom speech symbol dictionaries
-# Symbol dictionary files reside in the locale folder, e.g. `locale\en`, and are named `symbols-<name>.dic`.
-# If your add-on includes custom speech symbol dictionaries (most will not), fill out this dictionary.
-# Each key is the name of the dictionary,
-# with keys inside recording the following attributes:
-# displayName (name of the speech dictionary shown to users and translatable),
-# mandatory (True when always enabled, False when not.
+# Custom braille translation tables: display names
+brailleTableDisplayNames: dict[str, str] = {}
+
+# Custom symbol dictionaries
 symbolDictionaries: SymbolDictionaries = {}
+
+# Custom symbol dictionaries: display names
+symbolDictionaryDisplayNames: dict[str, str] = {}